save off todays changes.

+++ rserpool/include/rserpool.h
@@ -2,29 +2,29 @@
 #define __Rserpool_h__
 
 
-ssize_t rsp_socket();
+int rsp_socket();
 
-ssize_t rsp_close(int sockfd);
+int rsp_close(int sockfd);
 
-ssize_t rsp_connect(sockfd, name);
+int rsp_connect(sockfd, name);
 
-ssize_t rsp_register(sockfd, name);
+int rsp_register(sockfd, name);
 
-ssize_t rsp_deregister(sockfd, name);
+int rsp_deregister(sockfd, name);
 
 struct xxx rsp_getPoolInfo(sockfd, xxx );
 
-ssize_t rsp_reportfailure(sockfd, xxx);
+int rsp_reportfailure(sockfd, xxx);
 
-ssize_t rsp_sendmsg(int sockfd,         /* HA socket descriptor */
-                    struct msghdr *msg, /* message header struct */
-                    int flags);         /* Options flags */
+size_t rsp_sendmsg(int sockfd,         /* HA socket descriptor */
+                   struct msghdr *msg, /* message header struct */
+                   int flags);         /* Options flags */
 
 ssize_t rsp_rcvmsg(int sockfd,         /* HA socket descriptor */
                    struct msghdr *msg, /* msg header struct */
                    int flags);         /* Options flags */
 
 
-ssize_t rsp_forcefailover(sockfd, xxx);
+int rsp_forcefailover(sockfd, xxx);
 
 #endif
